supporting addition and removal from / to stdlib
Set correct update type for new items to be added and to be removed.
Also tweak filter checking for matching release update type to allow
item addition / removal for stdlib releases of type 'major'.

Last but not least, when a downgrade is deleted, also remove it from
the current array, not only the DB.