modify items/add to use PUT requests
Instead of requiring a POST of type multipart/form-data,
require a PUT of type application/x-ald-package. This is
more consistent to item descriptions / downloads and also
eliminates an old, unnecessary need to translate the data
to the old format.

Adjust the code accordingly: instead of the $_FILES array,
read the "php://input" stream in a temp file. Require the
Content-Length to be specified.