massive changes to FilterHelper to support more complex conditions
Pass the DB connection to the constructor. Split up into further
methods. For filters of type 'switch', allow a set of conditions
to be met. Validate comparison parameters. Add a new 'expr' type
and some structuring comments.

Use the modified FilterHelper class to re-implement the last pa-
rameter of the Suspension class. Adjust the filter code for can-
didate listing.